feat(community): add empty state and SEO
